requirements.txt
coverage==7.6.1
astunparse==1.6.3
langchain-openai==0.2.0
langchain-anthropic==0.2.1
langgraph==0.2.22
astor==0.8.1
python-dotenv==1.0.1
transformers==4.44.2
torch==2.4.1
datasets==3.0.0